Extracted all logic into components hook
